Accept a range of dates or time to query the audit log
This adds an (optional) argument 'after' which must be earlier in time than 'before'. Both arguments now accepts Date and POSIX* class objects instead of only a (numeric) UNIX timestamp (in milliseconds).

The 'time' column in the data frame with event details is now of class POSIXlt instead of a UNIX timestamp.